Fix formatting issue

+++ src/Day1-Data-types-and-Variables.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.28;
 
 contract DayOne {
     uint256 public postiveNumber = 42;
-    int public negativeNumber = -12;
+    int256 public negativeNumber = -12;
     string public text = "Day One solidity";
     bool public isTrue = true;
     address public owner;
